+++ src/main/java/net/minecraft/world/entity/ai/village/VillageSiege.java
@@ -116,7 +116,7 @@ public class VillageSiege implements CustomSpawner {
         for (int i = 0; i < 10; i++) {
             int x = pos.getX() + random.nextInt(16) - 8;
             int z = pos.getZ() + random.nextInt(16) - 8;
-            int y = level.getHeight(Heightmap.Types.WORLD_SURFACE, x, z);
+            long y = level.getHeight(Heightmap.Types.WORLD_SURFACE, x, z);
             BlockPos offset = new BlockPos(x, y, z);
             if (level.isVillage(offset) && Monster.checkMonsterSpawnRules(EntityTypes.ZOMBIE, level, EntitySpawnReason.EVENT, offset, random)) {
                 return Vec3.atBottomCenterOf(offset);
