+++ src/main/java/net/minecraft/world/level/chunk/ChunkSource.java
@@ -19,8 +19,9 @@ public abstract class ChunkSource implements AutoCloseable, LightChunkGetter {
     }
 
     @Override
-    public @Nullable LightChunk getChunkForLighting(final int x, final int z) {
-        return this.getChunk(x, z, ChunkStatus.EMPTY, false);
+    public @Nullable LightChunk getChunkForLighting(final long x, final long z) {
+        // MCRe NoiseFarlands: x/z 已 Long 化；getChunk 尚属区块管理模块（int 域），此处为 API 边界强转
+        return this.getChunk((int) x, (int) z, ChunkStatus.EMPTY, false);
     }
 
     public boolean hasChunk(final int x, final int z) {
@@ -31,7 +32,7 @@ public abstract class ChunkSource implements AutoCloseable, LightChunkGetter {
 
     public abstract void tick(BooleanSupplier haveTime, final boolean tickChunks);
 
-    public void onSectionEmptinessChanged(final int sectionX, final int sectionY, final int sectionZ, final boolean empty) {
+    public void onSectionEmptinessChanged(final long sectionX, final long sectionY, final long sectionZ, final boolean empty) {
     }
 
     public abstract String gatherStats();
