+++ src/main/java/net/minecraft/world/level/chunk/LightChunkGetter.java
@@ -6,7 +6,7 @@ import net.minecraft.world.level.LightLayer;
 import org.jspecify.annotations.Nullable;
 
 public interface LightChunkGetter {
-    @Nullable LightChunk getChunkForLighting(final int x, final int z);
+    @Nullable LightChunk getChunkForLighting(final long x, final long z);
 
     default void onLightUpdate(final LightLayer layer, final SectionPos pos) {
     }
